refactor: split risk period results
